fix: add support for symlinked files and dirs
This adds support for searching through symlinked files and directories,
the path is first canonicalized - which will resolve nested symlinks to
the real path. Due to the nature of symlinks it's still possible for an
infinitely recursive data structure to arise, or for the same canonical
file to be repeated. Both cases are covered by keeping track of which
canonical files have already been visited.

These changes are required to correctly support some Nix scenarios,
specifically installing fonts through home-manager.